Fix the param1 and param2 encoding
The values were all saved to `data[p]` but `p` was not updated, so only one param1 and param2 value were saved. This also led to invalid metadata decoding.
I have changed the function so that it keeps track of the size of `data` instead of inserting elements with `data[#data+1]` because the code is probably performance-relevant.